isolated pure locale handling functionality into LocaleUtils class to simplify Parser and JNLPFile
